+++ lib/braintree/dispute.rb
@@ -7,6 +7,7 @@ module Braintree
     attr_reader :amount_disputed
     attr_reader :amount_won
     attr_reader :case_number
+    attr_reader :chargeback_protection_level
     attr_reader :created_at
     attr_reader :currency_iso_code
     attr_reader :date_opened
@@ -17,6 +18,7 @@ module Braintree
     attr_reader :kind
     attr_reader :merchant_account_id
     attr_reader :original_dispute_id
+    attr_reader :paypal_messages
     attr_reader :processor_comments
     attr_reader :reason
     attr_reader :reason_code
@@ -61,6 +63,14 @@ module Braintree
       Chargeback = "chargeback"
       PreArbitration = "pre_arbitration"
       Retrieval = "retrieval"
+
+      All = constants.map { |c| const_get(c) }
+    end
+
+    module ChargebackProtectionLevel
+      Effortless = "effortless"
+      Standard = "standard"
+      NotProtected = "not_protected"
 
       All = constants.map { |c| const_get(c) }
     end
@@ -113,6 +123,10 @@ module Braintree
       @evidence = evidence.map do |record|
         Braintree::Dispute::Evidence.new(record)
       end unless evidence.nil?
+
+      @paypal_messages = paypal_messages.map do |record|
+        Braintree::Dispute::PayPalMessage.new(record)
+      end unless paypal_messages.nil?
 
       @transaction_details = TransactionDetails.new(transaction)
       @transaction = Transaction.new(transaction)