+++ spec/unit/braintree/dispute_search_spec.rb
@@ -41,6 +41,7 @@ describe Braintree::DisputeSearch do
   end
 
   [
+    :chargeback_protection_level,
     :kind,
     :reason,
     :status,